+++ crates/module-system/module-implementations/sov-sequencer-registry/src/lib.rs
@@ -122,6 +122,8 @@ impl<S: Spec> Module for SequencerRegistry<S> {
 
     fn genesis(
         &self,
+        _genesis_rollup_header: &<<S as Spec>::Da as DaSpec>::BlockHeader,
+        _validity_condition: &<<S as Spec>::Da as DaSpec>::ValidityCondition,
         config: &Self::Config,
         state: &mut impl GenesisState<S>,
     ) -> Result<(), Error> {