+++ crates/module-system/sov-modules-api/src/reexport_macros.rs
@@ -262,6 +262,8 @@ pub use sov_modules_macros::ModuleInfo;
 /// #
 /// #    fn genesis(
 /// #        &self,
+/// #        _genesis_slot_header: &<S::Da as DaSpec>::BlockHeader,
+/// #        _validity_condition: &<S::Da as DaSpec>::ValidityCondition,
 /// #        _config: &Self::Config,
 /// #        _state: &mut impl sov_modules_api::state::GenesisState<S>,
 /// #    ) -> Result<(), sov_modules_api::Error> {
@@ -309,6 +311,8 @@ pub use sov_modules_macros::ModuleInfo;
 /// #
 /// #    fn genesis(
 /// #        &self,
+/// #        _genesis_slot_header: &<S::Da as DaSpec>::BlockHeader,
+/// #        _validity_condition: &<S::Da as DaSpec>::ValidityCondition,
 /// #        _config: &Self::Config,
 /// #        _state: &mut impl sov_modules_api::state::GenesisState<S>,
 /// #    ) -> Result<(), sov_modules_api::Error> {
@@ -360,6 +364,8 @@ pub use sov_modules_macros::ModuleInfo;
 /// #
 /// #    fn genesis(
 /// #        &self,
+/// #        _genesis_slot_header: &<S::Da as DaSpec>::BlockHeader,
+/// #        _validity_condition: &<S::Da as DaSpec>::ValidityCondition,
 /// #        _config: &Self::Config,
 /// #        _state: &mut impl sov_modules_api::state::GenesisState<S>,
 /// #    ) -> Result<(), sov_modules_api::Error> {