+++ crates/module-system/sov-test-utils/src/runtime/mod.rs
@@ -300,7 +300,12 @@ where
                 .seq_da_address;
 
         let stf_state = storage_manager.create_storage();
-        let (state_root, change_set) = stf.init_chain(stf_state, genesis_config);
+        let (state_root, change_set) = stf.init_chain(
+            &Default::default(),
+            &Default::default(),
+            stf_state,
+            genesis_config,
+        );
 
         storage_manager.commit(change_set);